Add dummy-catalyst-mapping.json for mapping dummy catalyst items to actual catalyst items
This affects any of the older exotics that used to have "Upgrade Masterwork" and now have dummy items for the catalyst sockets.
Bungie uses reinitializationPossiblePlugHashes to autoapply the actual catalyst item once it's completed, but the values available to be socketed in do not include the actual catalyst item, so this mapping is needed for D2 tools to be able to automatically grab the real catalyst item based on the dummy item present in the socket options for a given exotic.